+++ foxcross/constants.py
@@ -1,0 +1,2 @@
+SLUGIFY_REGEX = r"([a-z](?=[A-Z])|[A-Z](?=[A-Z][a-z]))"
+SLUGIFY_REPLACE = r"\1-"