+++ foxcross/enums.py
@@ -3,5 +3,7 @@ from enum import Enum
 
 class MediaTypes(Enum):
     JSON = "application/json"
+    HTML = "text/html"
+    ANY_TEXT = "text/*"
     ANY_APP = "application/*"
     ANY = "*/*"