+++ foxcross/pandas_serving.py
@@ -1,5 +1,5 @@
 import logging
-from typing import Dict, Union
+from typing import Any, Dict, Union
 
 from starlette.exceptions import HTTPException
 
@@ -18,14 +18,6 @@ except ImportError:
             f"Cannot import pandas. Please install foxcross using foxcross[pandas] or"
             f" foxcross[modin]"
         )
-
-try:
-    import ujson as json
-    from starlette.responses import UJSONResponse as JSONResponse
-except ImportError:
-    import json  # noqa: F401
-    from starlette.responses import JSONResponse
-
 
 logger = logging.getLogger(__name__)
 
@@ -72,23 +64,28 @@ class DataFrameModelServing(ModelServing):
 
     def _format_output(
         self, results: Union[pandas.DataFrame, Dict[str, pandas.DataFrame]]
-    ) -> JSONResponse:
+    ) -> Any:
         # Convert NaNs to Nones to handle ujson OverflowError
         try:
-            results = results.replace({numpy.nan: None}).to_dict(
-                orient=self.pandas_orient
-            )
+            output = results.replace({numpy.nan: None}).to_dict(orient=self.pandas_orient)
         except AttributeError:
-            results = {
-                key: value.replace({numpy.nan: None}).to_dict(orient=self.pandas_orient)
-                for key, value in results.items()
-            }
-            results["multi_dataframe"] = True
-        return self._get_response(results)
+            try:
+                output = {
+                    key: value.replace({numpy.nan: None}).to_dict(
+                        orient=self.pandas_orient
+                    )
+                    for key, value in results.items()
+                }
+                output["multi_dataframe"] = True
+            except (TypeError, AttributeError):
+                err_msg = f"Failed to format prediction results"
+                logger.exception(err_msg)
+                raise HTTPException(status_code=500, detail=err_msg)
+        return output
 
 
 _model_serving_runner = ModelServingRunner(
-    ModelServing, [ModelServing, DataFrameModelServing]
+    ModelServing, (ModelServing, DataFrameModelServing)
 )
 compose_pandas = _model_serving_runner.compose
 run_pandas_serving = _model_serving_runner.run_model_serving