+++ tests/test_pandas_serving.py
@@ -1,4 +1,5 @@
 import os
+import re
 from pathlib import Path
 from typing import Dict, Union
 
@@ -6,6 +7,7 @@ import pytest
 from slugify import slugify
 from starlette.testclient import TestClient
 
+from foxcross.constants import SLUGIFY_REGEX, SLUGIFY_REPLACE
 from foxcross.enums import MediaTypes
 from foxcross.pandas_serving import DataFrameModelServing, compose_pandas
 
@@ -49,6 +51,15 @@ with Path(__location__ / "data/interpolate_multi_frame_result.json").open() as f
 
 class PredictMethodNotDefined(DataFrameModelServing):
     test_data_path = interpolate_data_path
+
+
+class FormatOutputAttributeError(DataFrameModelServing):
+    test_data_path = interpolate_data_path
+
+    def predict(
+        self, data: Union[pandas.DataFrame, Dict[str, pandas.DataFrame]]
+    ) -> Union[pandas.DataFrame, Dict[str, pandas.DataFrame]]:
+        return {"hi": "there", "my": "friend"}
 
 
 class InterpolateModel:
@@ -137,16 +148,22 @@ def test_index_single_model_serving():
 def test_predict_multi_model_serving():
     app = compose_pandas(__name__, debug=True)
     client = TestClient(app)
+    interp_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, InterpolateModelServing.__name__)
+    )
     response = client.post(
-        f"{slugify(InterpolateModelServing.__name__)}/predict/",
+        f"{interp_slugified}/predict/",
         headers={"Accept": MediaTypes.JSON.value},
         json=interpolate_data,
     )
     assert response.status_code == 200
     assert response.json() == interpolate_result_data
 
+    multi_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, InterpolateMultiFrameModelServing.__name__)
+    )
     multi_response = client.post(
-        f"{slugify(InterpolateMultiFrameModelServing.__name__)}/predict/",
+        f"{multi_slugified}/predict/",
         headers={"Accept": MediaTypes.JSON.value},
         json=interpolate_multi_frame_data,
     )
@@ -157,10 +174,17 @@ def test_predict_multi_model_serving():
 def test_index_multi_model_serving():
     app = compose_pandas(__name__, debug=True)
     client = TestClient(app)
-    response = client.get(f"{slugify(InterpolateModelServing.__name__)}/")
+
+    interp_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, InterpolateModelServing.__name__)
+    )
+    response = client.get(f"{interp_slugified}/")
     assert response.status_code == 200
 
-    multi_response = client.get(f"{slugify(InterpolateMultiFrameModelServing.__name__)}/")
+    multi_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, InterpolateMultiFrameModelServing.__name__)
+    )
+    multi_response = client.get(f"{multi_slugified}/")
     assert multi_response.status_code == 200
 
     root_response = client.get("/")
@@ -184,23 +208,30 @@ def test_endpoints_multi_model_serving(
 ):
     app = compose_pandas(__name__, debug=True)
     client = TestClient(app)
+
+    interp_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, InterpolateModelServing.__name__)
+    )
     response = client.get(
-        f"{slugify(InterpolateModelServing.__name__)}{endpoint}",
-        headers={"Accept": MediaTypes.JSON.value},
+        f"{interp_slugified}{endpoint}", headers={"Accept": MediaTypes.JSON.value}
     )
     assert response.status_code == 200
     assert response.json() == first_expected
 
+    multi_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, InterpolateMultiFrameModelServing.__name__)
+    )
     multi_response = client.get(
-        f"{slugify(InterpolateMultiFrameModelServing.__name__)}{endpoint}",
-        headers={"Accept": MediaTypes.JSON.value},
+        f"{multi_slugified}{endpoint}", headers={"Accept": MediaTypes.JSON.value}
     )
     assert multi_response.status_code == 200
     assert multi_response.json() == second_expected
 
+    add_one_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, AddOneModel.__name__)
+    )
     add_one_response = client.get(
-        f"{slugify(AddOneModel.__name__)}{endpoint}",
-        headers={"Accept": MediaTypes.JSON.value},
+        f"{add_one_slugified}{endpoint}", headers={"Accept": MediaTypes.JSON.value}
     )
     assert add_one_response.status_code == 200
     assert add_one_response.json() == third_expected
@@ -224,3 +255,48 @@ def test_predict_input_issue():
         "/predict/", headers={"Accept": MediaTypes.JSON.value}, json=[]
     )
     assert multi_response.status_code == 400
+
+
+@pytest.mark.parametrize("endpoint", ["/predict/", "/predict-test/", "/input-format/"])
+def test_single_model_html_responses(endpoint):
+    app = InterpolateMultiFrameModelServing(debug=True)
+    client = TestClient(app)
+    response = client.get(endpoint, headers={"Accept": MediaTypes.HTML.value})
+    assert response.status_code == 200
+
+
+@pytest.mark.parametrize("endpoint", ["/predict/", "/predict-test/", "/input-format/"])
+def test_multi_model_html_responses(endpoint):
+    app = compose_pandas(__name__, debug=True)
+    client = TestClient(app)
+
+    interp_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, InterpolateModelServing.__name__)
+    )
+    response = client.get(
+        f"{interp_slugified}{endpoint}", headers={"Accept": MediaTypes.HTML.value}
+    )
+    assert response.status_code == 200
+
+    multi_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, InterpolateMultiFrameModelServing.__name__)
+    )
+    multi_response = client.get(
+        f"{multi_slugified}{endpoint}", headers={"Accept": MediaTypes.HTML.value}
+    )
+    assert multi_response.status_code == 200
+
+    add_one_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, AddOneModel.__name__)
+    )
+    add_one_response = client.get(
+        f"{add_one_slugified}{endpoint}", headers={"Accept": MediaTypes.HTML.value}
+    )
+    assert add_one_response.status_code == 200
+
+
+def test_format_output_attribute_error():
+    app = FormatOutputAttributeError(debug=True)
+    client = TestClient(app)
+    response = client.get("/predict-test/")
+    assert response.status_code == 500